Renamed transcription and transcriber endpoint file
